Move $nu creation from engine to protocol
This adds nu-path and sysinfo dependencies to nu-protocol, but removes
sysinfo from nu-engine.
