+++ crates/nu-protocol/Cargo.toml
@@ -14,6 +14,7 @@ bench = false
 
 [dependencies]
 nu-utils = { path = "../nu-utils", version = "0.84.1" }
+nu-path = { path = "../nu-path", version = "0.84.1" }
 
 byte-unit = "4.0"
 chrono = { version = "0.4", features = [ "serde", "std", "unstable-locales" ], default-features = false }
@@ -25,6 +26,7 @@ miette = { version = "5.10", features = ["fancy-no-backtrace"] }
 num-format = "0.4"
 serde = { version = "1.0", default-features = false }
 serde_json = { version = "1.0", optional = true }
+sysinfo = "0.29"  # required for creating $nu
 thiserror = "1.0"
 typetag = "0.2"
